card_organizer/requirements.txt
asgiref==3.2.5
coverage==5.0.4
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
mysqlclient==1.4.6
pokemontcgsdk==2.0.0
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
sqlparse==0.3.1
